requirements.txt
python-binance==1.0.16
python-dotenv==1.0.0
schedule==1.1.0
requests==2.31.0
langchain-community==0.0.29
langchain-openai==0.0.8
langchain-experimental==0.0.55
python-dotenv==1.0.0
pandas==2.2.1
numpy==1.26.4
openai==1.12.0
